refactor: test dirs
